+++ logic_assign.lua
@@ -1,0 +1,9 @@
+local Logic = BuffDuty.Logic
+
+------------ ASSIGN LOGIC ------------
+local assignLogic = {}
+Logic[Logic.Type.ASSIGN] = assignLogic
+
+function assignLogic.generateDutyMap(cmd, raid_info, class_players)
+    -- TODO
+end